2주차 과제 2 - 간단한 Todo App 만들기
- 피드백 수정 (변수명 변경, 컴포넌트 분리)
